\name{kmBlockOld-package}
\alias{kmBlockOld-package}
\alias{kmBlockOld}
\docType{package}
\title{
\packageTitle{kmBlockOld}
}
\description{
\packageDescription{kmBlockOld}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{kmBlockOld}
\packageIndices{kmBlockOld}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{kmBlockOld}

Maintainer: \packageMaintainer{kmBlockOld}
}
\references{
~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
